requirements.txt
PyNaCl>=1.5.0
requests>=2.28.0
websocket-client>=1.5.0
numpy>=1.24.0
python-dotenv>=1.0.0
starknet-py>=0.18.0
Flask>=2.3.0
flask-socketio>=5.3.0
python-socketio>=5.9.0
werkzeug>=2.3.0
cryptography>=41.0.0